+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ This project adheres to `Semantic Versioning`_ starting with version 0.7.0.
 
 Added
 -----
+- Returning used model name and project name in the response of ``GET /parse`` and ``POST /parse`` as ``model`` and ``project`` respectively.
 
 Changed
 -------
@@ -22,7 +23,7 @@ Fixed
 -----
 - readded possibility to set fixed model name from http train endpoint
 
-[0.12.2] - 2018-03-20
+[0.12.2] - 2018-04-20
 ^^^^^^^^^^^^^^^^^^^^^
 
 Fixed
@@ -30,8 +31,11 @@ Fixed
 - fixed duckling text extraction for ner_duckling_http 
 
 
-[0.12.1] - 2018-03-18
+[0.12.1] - 2018-04-18
 ^^^^^^^^^^^^^^^^^^^^^
+Added
+-----
+- support for retrieving training data from a URL
 
 Fixed
 -----
@@ -40,7 +44,7 @@ Fixed
   documentation
 
 
-[0.12.0] - 2018-03-17
+[0.12.0] - 2018-04-17
 ^^^^^^^^^^^^^^^^^^^^^
 
 Added